requirements.txt
Django==4.2.8
psycopg2>=2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers
drf-spectacular
drf-spectacular-sidecar